+++ src/main/java/com/github/games647/changeskin/ChangeSkin.java
@@ -4,22 +4,15 @@ import com.comphenix.protocol.utility.SafeCacheBuilder;
 import com.comphenix.protocol.wrappers.WrappedSignedProperty;
 import com.github.games647.changeskin.listener.AsyncPlayerLoginListener;
 import com.github.games647.changeskin.listener.PlayerLoginListener;
-import com.github.games647.changeskin.tasks.SkinDownloader;
-import com.google.common.base.Charsets;
 import com.google.common.cache.CacheLoader;
 import com.google.common.collect.Lists;
-import com.google.common.collect.Maps;
-import com.google.common.io.Files;
 
 import java.io.BufferedReader;
-import java.io.BufferedWriter;
-import java.io.File;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.net.HttpURLConnection;
 import java.net.URL;
 import java.util.List;
-import java.util.Map;
 import java.util.UUID;
 import java.util.concurrent.ConcurrentMap;
 import java.util.concurrent.TimeUnit;
@@ -38,21 +31,8 @@ public class ChangeSkin extends JavaPlugin {
     private static final String UUID_URL = "https://api.mojang.com/users/profiles/minecraft/";
 
     private final List<SkinData> defaultSkins = Lists.newArrayList();
-    private final ConcurrentMap<UUID, UUID> userPreferences = Maps.newConcurrentMap();
 
-    //this is thread-safe in order to save and load from different threads like the skin download
-    private final ConcurrentMap<UUID, WrappedSignedProperty> skinCache = SafeCacheBuilder
-            .<UUID, WrappedSignedProperty>newBuilder()
-            .maximumSize(1024 * 5)
-            .expireAfterWrite(5, TimeUnit.HOURS)
-            .build(new CacheLoader<UUID, WrappedSignedProperty>() {
-
-                @Override
-                public WrappedSignedProperty load(UUID key) throws Exception {
-                    //A key should be inserted manually on start packet
-                    throw new UnsupportedOperationException("Not supported");
-                }
-            });
+    private Storage storage;
 
     //this is thread-safe in order to save and load from different threads like the skin download
     private final ConcurrentMap<String, UUID> uuidCache = SafeCacheBuilder
@@ -72,8 +52,16 @@ public class ChangeSkin extends JavaPlugin {
     public void onEnable() {
         saveDefaultConfig();
 
+        storage = new Storage(this);
+        try {
+            storage.createTables();
+        } catch (Exception ex) {
+            getLogger().log(Level.SEVERE, "Failed to setup database. Disabling plugin...", ex);
+            setEnabled(false);
+            return;
+        }
+
         loadDefaultSkins();
-        loadPreferences();
 
         getCommand("setskin").setExecutor(new SetSkinCommand(this));
 
@@ -83,21 +71,17 @@ public class ChangeSkin extends JavaPlugin {
 
     @Override
     public void onDisable() {
-        savePreferences();
-
         //clean up
-        userPreferences.clear();
-        skinCache.clear();
+        if (storage != null) {
+            storage.close();
+        }
+
         defaultSkins.clear();
         uuidCache.clear();
     }
 
-    public ConcurrentMap<UUID, WrappedSignedProperty> getSkinCache() {
-        return skinCache;
-    }
-
-    public ConcurrentMap<UUID, UUID> getUserPreferences() {
-        return userPreferences;
+    public Storage getStorage() {
+        return storage;
     }
 
     public List<SkinData> getDefaultSkins() {
@@ -108,7 +92,7 @@ public class ChangeSkin extends JavaPlugin {
         return uuidCache;
     }
 
-    public WrappedSignedProperty downloadSkin(UUID ownerUUID) {
+    public SkinData downloadSkin(UUID ownerUUID) {
         //unsigned is needed in order to receive the signature
         String uuidString = ownerUUID.toString().replace("-", "") + "?unsigned=false";
         try {
@@ -121,13 +105,16 @@ public class ChangeSkin extends JavaPlugin {
                 JSONObject userData = (JSONObject) JSONValue.parseWithException(line);
 
                 JSONArray properties = (JSONArray) userData.get("properties");
-                JSONObject skinData = (JSONObject) properties.get(0);
+                JSONObject data = (JSONObject) properties.get(0);
 
                 //base64 encoded skin data
-                String encodedSkin = (String) skinData.get("value");
-                String signature = (String) skinData.get("signature");
+                String encodedSkin = (String) data.get("value");
+                String signature = (String) data.get("signature");
 
-                return WrappedSignedProperty.fromValues("textures", encodedSkin, signature);
+                WrappedSignedProperty property = WrappedSignedProperty.fromValues("textures", encodedSkin, signature);
+                SkinData skinData = new SkinData(property.getValue(), property.getSignature());
+
+                return skinData;
             }
         } catch (IOException ioException) {
             getLogger().log(Level.SEVERE, "Tried downloading skin data from Mojang", ioException);
@@ -159,15 +146,37 @@ public class ChangeSkin extends JavaPlugin {
         return null;
     }
 
-    public void setSkin(Player player, UUID targetSkin) {
-        userPreferences.put(player.getUniqueId(), targetSkin);
-        if (!skinCache.containsKey(targetSkin)) {
-            //download the skin only if it's not already in the cache
-            getServer().getScheduler().runTaskAsynchronously(this, new SkinDownloader(this, null, player, targetSkin));
-        }
+    //you should call this method async
+    public void setSkin(Player player, SkinData newSkin, boolean applyNow) {
+        final UserPreferences preferences = storage.getPreferences(player.getUniqueId(), false);
+        preferences.setTargetSkin(newSkin);
+        getServer().getScheduler().runTaskAsynchronously(this, new Runnable() {
+            @Override
+            public void run() {
+                storage.save(preferences);
+            }
+        });
     }
 
-    private UUID parseId(String withoutDashes) {
+    //you should call this method async
+    public void setSkin(Player player, UUID targetSkin, boolean applyNow) {
+        SkinData newSkin = storage.getSkin(targetSkin, true);
+        if (newSkin == null) {
+            newSkin = downloadSkin(targetSkin);
+
+            final SkinData skin = newSkin;
+            getServer().getScheduler().runTaskAsynchronously(this, new Runnable() {
+                @Override
+                public void run() {
+                    storage.save(skin);
+                }
+            });
+        }
+
+        setSkin(player, newSkin, applyNow);
+    }
+
+    public static UUID parseId(String withoutDashes) {
         return UUID.fromString(withoutDashes.substring(0, 8)
                 + "-" + withoutDashes.substring(8, 12)
                 + "-" + withoutDashes.substring(12, 16)
@@ -175,81 +184,13 @@ public class ChangeSkin extends JavaPlugin {
                 + "-" + withoutDashes.substring(20, 32));
     }
 
-    private void savePreferences() {
-        getDataFolder().mkdir();
-        File file = new File(getDataFolder(), "preferences.txt");
-
-        BufferedWriter bufferedWriter = null;
-        try {
-            if (!file.exists()) {
-                file.createNewFile();
-            }
-
-            bufferedWriter = Files.newWriter(file, Charsets.UTF_8);
-            for (Map.Entry<UUID, UUID> entry : userPreferences.entrySet()) {
-                if (entry.getKey().equals(entry.getValue())) {
-                    continue;
-                }
-
-                bufferedWriter.write(entry.getKey().toString());
-                bufferedWriter.write(':');
-                bufferedWriter.write(entry.getValue().toString());
-                bufferedWriter.newLine();
-            }
-
-            bufferedWriter.flush();
-        } catch (IOException ioExc) {
-            getLogger().log(Level.SEVERE, "Failed to save skin prefernces", ioExc);
-        } finally {
-            if (bufferedWriter != null) {
-                try {
-                    bufferedWriter.close();
-                } catch (IOException ioExc) {
-                    getLogger().log(Level.SEVERE, "Failed to close the file handle", ioExc);
-                }
-            }
-        }
-    }
-
-    private void loadPreferences() {
-        getDataFolder().mkdir();
-        File file = new File(getDataFolder(), "preferences.txt");
-
-        BufferedReader bufferedReader = null;
-        try {
-            if (!file.exists()) {
-                file.createNewFile();
-            }
-
-            bufferedReader = Files.newReader(file, Charsets.UTF_8);
-            String currentLine = bufferedReader.readLine();
-            while (currentLine != null && !currentLine.isEmpty()) {
-                String[] parts = currentLine.split(":");
-                UUID player = UUID.fromString(parts[0]);
-                UUID target = UUID.fromString(parts[1]);
-                userPreferences.put(player, target);
-
-                currentLine = bufferedReader.readLine();
-            }
-        } catch (IOException ioExc) {
-            getLogger().log(Level.SEVERE, "Failed to load preferences", ioExc);
-        } finally {
-            if (bufferedReader != null) {
-                try {
-                    bufferedReader.close();
-                } catch (IOException ioExc) {
-                    getLogger().log(Level.SEVERE, "Failed to close the file handle", ioExc);
-                }
-            }
-        }
-    }
-
     private void loadDefaultSkins() {
         List<String> defaultList = getConfig().getStringList("default-skins");
         for (String uuidString : defaultList) {
             UUID ownerUUID = UUID.fromString(uuidString);
-            WrappedSignedProperty skinData = downloadSkin(ownerUUID);
-            defaultSkins.add(new SkinData(skinData, ownerUUID));
+
+            SkinData skinData = downloadSkin(ownerUUID);
+            defaultSkins.add(skinData);
         }
     }
 }