+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ##### 0.8
 
+* Added database storage
 * Added reset command /skin reset
 * Added specific skin permssions. You can use changeskin.skin.* changeskin.skin.playerName changeskin.skin.uuid
 * Fixed invisible players will be make visible